end of section 21 admin

+++ client/src/app/core/services/cart.service.ts
@@ -128,12 +128,14 @@ export class CartService {
   }
 
   deleteCart() {
-    this.http.delete(this.baseUrl + 'cart?id=' + this.cart()?.id).subscribe({
-      next: () => {
-        localStorage.removeItem('cart_id');
-        this.cart.set(null);
-        this.selectedDelivery.set(null);
-      }
-    })
-  }  
+    if (this.cart()) {
+      this.http.delete(this.baseUrl + 'cart?id=' + this.cart()?.id).subscribe({
+        next: () => {
+          localStorage.removeItem('cart_id');
+          this.cart.set(null);
+          this.selectedDelivery.set(null);
+        }
+      })
+    }
+  }
 }
